Update some testcases now that 64-bit immediates are working.

+++ tests/mini/naked_asm2.d
@@ -1,6 +1,6 @@
 int foo()
 {
-    static int fourty2 = 42;
+    static size_t fourty2 = 42;
     version(X86)
     asm
     {
